# Make -delta-test=none default for benchstat in funcbench